+++ api/configuration/v1alpha1/kongcertificate_types.go
@@ -1,10 +1,21 @@
 package v1alpha1
 
 import (
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	commonv1alpha1 "github.com/kong/kong-operator/api/common/v1alpha1"
 	konnectv1alpha2 "github.com/kong/kong-operator/api/konnect/v1alpha2"
+)
+
+// KongCertificateSourceType is the type of source for the certificate data.
+type KongCertificateSourceType string
+
+const (
+	// KongCertificateSourceTypeInline indicates that the certificate data is provided inline in the spec.
+	KongCertificateSourceTypeInline KongCertificateSourceType = "inline"
+	// KongCertificateSourceTypeSecretRef indicates that the certificate data is sourced from a Kubernetes Secret.
+	KongCertificateSourceTypeSecretRef KongCertificateSourceType = "secretRef"
 )
 
 // KongCertificate is the schema for Certificate API which defines a Kong Certificate.
@@ -36,8 +47,20 @@ type KongCertificate struct {
 // +kubebuilder:validation:XValidation:rule="!has(self.controlPlaneRef) ? true : self.controlPlaneRef.type != 'kic'", message="KIC is not supported as control plane"
 // +kubebuilder:validation:XValidation:rule="!has(self.adopt) ? true : (has(self.controlPlaneRef) && self.controlPlaneRef.type == 'konnectNamespacedRef')", message="spec.adopt is allowed only when controlPlaneRef is konnectNamespacedRef"
 // +kubebuilder:validation:XValidation:rule="(has(oldSelf.adopt) && has(self.adopt)) || (!has(oldSelf.adopt) && !has(self.adopt))", message="Cannot set or unset spec.adopt in updates"
+// +kubebuilder:validation:XValidation:rule="self.type == 'inline' ? (has(self.cert) && self.cert.size() != 0 && has(self.key) && self.key.size() != 0) : true", message="spec.cert and spec.key are required when type is 'inline'"
+// +kubebuilder:validation:XValidation:rule="self.type == 'secretRef' ? has(self.secretRef) : true", message="spec.secretRef is required when type is 'secretRef'"
+// +kubebuilder:validation:XValidation:rule="!((has(self.cert) || has(self.key)) && (has(self.secretRef) || has(self.secretRefAlt)))", message="cert/key and secretRef/secretRefAlt cannot be set at the same time"
+// +kubebuilder:validation:XValidation:rule="!((has(self.cert_alt) || has(self.key_alt)) && (has(self.secretRef) || has(self.secretRefAlt)))", message="cert_alt/key_alt and secretRef/secretRefAlt cannot be set at the same time"
 // +apireference:kgo:include
 type KongCertificateSpec struct {
+	// Type indicates the source of the certificate data.
+	// Can be 'inline' or 'secretRef'.
+	// +required
+	// +kubebuilder:validation:Enum=inline;secretRef
+	// +kubebuilder:default=inline
+	// +optional
+	Type *KongCertificateSourceType `json:"type,omitempty"`
+
 	// ControlPlaneRef references the Konnect Control Plane that this KongCertificate should be created in.
 	// +kubebuilder:validation:XValidation:message="'konnectID' type is not supported", rule="self.type != 'konnectID'"
 	// +required
@@ -47,6 +70,20 @@ type KongCertificateSpec struct {
 	// +optional
 	Adopt *commonv1alpha1.AdoptOptions `json:"adopt,omitempty"`
 
+	// SecretRef is a reference to a Kubernetes Secret containing the certificate and key.
+	// This field is used when type is 'secretRef'.
+	// The Secret must contain keys named 'tls.crt' and 'tls.key'.
+	// +optional
+	SecretRef *corev1.SecretReference `json:"secretRef,omitempty"`
+
+	// SecretRefAlt is a reference to a Kubernetes Secret containing the alternative certificate and key.
+	// This should only be set if you have both RSA and ECDSA types of certificate available
+	// and would like Kong to prefer serving using ECDSA certs when client advertises support for it.
+	// This field is used when type is 'secretRef'.
+	// The Secret must contain keys named 'tls.crt' and 'tls.key'.
+	// +optional
+	SecretRefAlt *corev1.SecretReference `json:"secretRefAlt,omitempty"`
+
 	KongCertificateAPISpec `json:",inline"`
 }
 
@@ -54,20 +91,26 @@ type KongCertificateSpec struct {
 // +apireference:kgo:include
 type KongCertificateAPISpec struct {
 	// Cert is the PEM-encoded certificate.
-	// +required
+	// This field is used when type is 'inline'.
+	// +optional
 	Cert string `json:"cert,omitempty"`
 	// CertAlt is the PEM-encoded certificate.
 	// This should only be set if you have both RSA and ECDSA types of
 	// certificate available and would like Kong to prefer serving using ECDSA certs
 	// when client advertises support for it.
+	// This field is used when type is 'inline'.
+	// +optional
 	CertAlt string `json:"cert_alt,omitempty"`
 	// Key is the PEM-encoded private key.
-	// +required
+	// This field is used when type is 'inline'.
+	// +optional
 	Key string `json:"key,omitempty"`
 	// KeyAlt is the PEM-encoded private key.
 	// This should only be set if you have both RSA and ECDSA types of
 	// certificate available and would like Kong to prefer serving using ECDSA certs
 	// when client advertises support for it.
+	// This field is used when type is 'inline'.
+	// +optional
 	KeyAlt string `json:"key_alt,omitempty"`
 
 	// Tags is an optional set of tags applied to the certificate.