+++ internal/utils/index/kongcertificate.go
@@ -9,6 +9,8 @@ import (
 const (
 	// IndexFieldKongCertificateOnKonnectGatewayControlPlane is the index field for KongCertificate -> KonnectGatewayControlPlane.
 	IndexFieldKongCertificateOnKonnectGatewayControlPlane = "kongCertificateKonnectGatewayControlPlaneRef"
+	// IndexFieldKongCertificateReferencesSecrets is the index field for KongCertificate -> Secret.
+	IndexFieldKongCertificateReferencesSecrets = "kongCertificateSecretRef"
 )
 
 // OptionsForKongCertificate returns required Index options for KongCertificate reconciler.
@@ -19,5 +21,24 @@ func OptionsForKongCertificate(cl client.Client) []Option {
 			Field:          IndexFieldKongCertificateOnKonnectGatewayControlPlane,
 			ExtractValueFn: indexKonnectGatewayControlPlaneRef[configurationv1alpha1.KongCertificate](cl),
 		},
+		{
+			Object:         &configurationv1alpha1.KongCertificate{},
+			Field:          IndexFieldKongCertificateReferencesSecrets,
+			ExtractValueFn: secretOnKongCertificate,
+		},
 	}
+}
+
+// secretOnKongCertificate indexes KongCertificate by its referenced Secret.
+func secretOnKongCertificate(object client.Object) []string {
+	cert, ok := object.(*configurationv1alpha1.KongCertificate)
+	if !ok {
+		return nil
+	}
+
+	if cert.Spec.SecretRef == nil {
+		return nil
+	}
+
+	return []string{cert.Spec.SecretRef.Namespace + "/" + cert.Spec.SecretRef.Name}
 }