+++ CHANGELOG.md
@@ -43,6 +43,10 @@
 - DataPlane: Enable incremental config sync by default when using Konnect as control plane.
   This improves performance of config syncs for large configurations.
   [#2759](https://github.com/Kong/kong-operator/pull/2759)
+- KongCertificate: Add support for sourcing certificates from Kubernetes Secrets.                                                                                                      │
+  This allows users to define KongCertificates that reference existing Kubernetes                                                                                                      │
+  Secrets containing TLS certificate and key data, instead of embedding them inline.                                                                                                   │
+  [#2802](https://github.com/Kong/kong-operator/pull/2802)  
 
 ### Changed